lsd: remove GNU releases, package binary unpacked
Closes https://github.com/Starz0r/ChocolateyPackagingScripts/issues/22

+++ src/templates/lsd/tools/chocolateyInstall.ps1
@@ -2,25 +2,11 @@ $PackageName = 'lsd';
 $ToolsPath   = Split-Path -Parent $MyInvocation.MyCommand.Definition;
 $PkgParams   = Get-PackageParameters;
 
-If ($PkgParams.GNU) {
-	$PackageArgs = @{
-		PackageName    = $PackageName
-		FileFullPath   = Get-Item $(Join-Path $ToolsPath '$altfname')
-		FileFullPath64 = Get-Item $(Join-Path $ToolsPath '$altfname64')
-		Destination    = $ToolsPath
-	};
-} Else {
-	$PackageArgs = @{
-		PackageName    = $PackageName
-		FileFullPath   = Get-Item $(Join-Path $ToolsPath '$fname')
-		FileFullPath64 = Get-Item $(Join-Path $ToolsPath '$fname64')
-		Destination    = $ToolsPath
-	};
-};
-
 # Remove Previous Installations
 Get-ChildItem $ToolsPath\* | Where-Object { $_.PSISContainer } | ForEach-Object {
-	Remove-Item $_ -Recurse -Force -EA 0 | Out-Null;
-};
-Get-ChocolateyUnzip @PackageArgs;
+	IF (($_.BaseName -ne "x86") -and ($_.BaseName -ne "x64")) {
+		Write-Warning "Deleting $_ from previous deprecated installation process. Chocolatey may continue to refer to this path until you uninstall and reinstall this package. No further action is required.";
+		Remove-Item $_ -Recurse -Force -EA 0 | Out-Null;
+	}
+}
 Remove-Item ($ToolsPath + '\*.' + 'zip');
